chatsphere_agent/requirements.txt
langchain
langchain-google-genai
python-dotenv
pinecone
psycopg2-binary
langchain-community
